// packages/codex/src/protocol/generated/ClientNotification.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`.
// Generated from codex-cli 0.153.4, licensed Apache-2.0.
// Regenerate: `bun run generate:protocol` from packages/codex.
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ClientNotification = { "method": "initialized" };